backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "2D to 3D Floorplan ML Backend"
readme = "README.md"

# IMPORTANT: Python version supported by PyTorch + FastAPI
requires-python = ">=3.11,<3.12"

dependencies = [
    "fastapi",
    "uvicorn",
    "numpy",
    "opencv-python",
    "pillow",
    "scikit-image",
    "shapely",
    "trimesh",
    "python-multipart",
    "torch",
    "torchvision",
]

